feat: Adding existence checks.

+++ app/Services/ClassService.ts
@@ -9,9 +9,25 @@ export default class ClassService {
   private async checkExists (id_class: number) {
     return await this.repository.find(id_class)
   }
+  private async checkStudentExists (id_student: number) {
+    return await userService.showUser(id_student)
+  }
   public async getClasses () {
     try {
       const classes = await this.repository.getAll()
+      return classes
+    } catch (error) {
+      throw new Error(error.message)
+    }
+  }
+  public async getClassesbyStudent (id_student: number) {
+    try {
+      const student = await this.checkStudentExists(id_student)
+
+      if (!student) {
+        throw new Error('Student not found')
+      }
+      const classes = await this.repository.getByStudent(student)
       return classes
     } catch (error) {
       throw new Error(error.message)
@@ -39,21 +55,29 @@ export default class ClassService {
   }
   private async studentExists (id_student: number, id_class: number) {
     try {
-      return await this.repository.studentExists(id_student, id_class)
+      const exists = await this.repository.studentExists(id_student, id_class)
+      return exists
     } catch (error) {
       throw new Error(error.message)
     }
   }
-  public async addStudent (id_class: number, id_student: number) {
+  public async addStudent (id_class: number, id_student: number, id_teacher: number) {
     try {
       const student = await userService.showUser(id_student)
       if (!student) {
         throw new Error('Student not found')
       }
+
       const clss = await this.checkExists(id_class)
 
       if (!clss) {
         throw new Error('Class not found')
+      }
+
+      const teacher = id_teacher === clss.id_teacher
+
+      if (!teacher) {
+        throw new Error('Not permited!')
       }
 
       const full = !clss.available
@@ -64,7 +88,7 @@ export default class ClassService {
 
       const exists = await this.studentExists(id_student, id_class)
 
-      if (!exists) {
+      if (exists) {
         throw new Error('Student is already in this class!')
       }
 
@@ -78,14 +102,25 @@ export default class ClassService {
     }
   }
   public async getStudents (id_class: number) {
-    try{
-      const students = this.repository.getStudents(id_class)
-    }catch(error){
+    try {
+      const clss = await this.repository.find(id_class)
+
+      if (!clss) {
+        throw new Error('Class not found!')
+      }
+      const students = this.repository.getStudents(clss)
+      return students
+    } catch (error) {
       throw new Error(error.message)
     }
   }
   public async removeStudent (id_class: number, id_student: number) {
     try {
+      const exists = await this.studentExists(id_student, id_class)
+      console.log(exists)
+      if (!exists) {
+        throw new Error('Could not find this student in this class')
+      }
       return await this.repository.deleteStudent(id_student, id_class)
     } catch (error) {
       throw new Error(error.message)
@@ -104,9 +139,24 @@ export default class ClassService {
       throw new Error(error.message)
     }
   }
+  public async editClass (id_class: number, data: object) {
+    try {
+      const exists = await this.checkExists(id_class)
+
+      if (!exists) {
+        throw new Error('Class not found')
+      }
+      const clss = await this.repository.edit(id_class, data)
+      return clss
+    } catch (error) {
+      throw new Error(error.message)
+    }
+  }
   public async removeClass (id_class: number) {
     try {
       const exists = await this.checkExists(id_class)
+
+      console.log(exists)
 
       if (!exists) {
         throw new Error('Class not found!')
@@ -115,7 +165,7 @@ export default class ClassService {
 
       return Class
     } catch (error) {
-      return new Error(error.message)
+      throw new Error(error.message)
     }
   }
 }
